Breaking out SAML and generic Social Auth
